Add an 'out_dir' option to jsonnet_to_json()
If jsonnet_to_json() yields a large number of output files, or yields a
set of paths that is not known up front, it would be preferable to store
these in a directory artifact (i.e., created using
ctx.actions.declare_directory()).

This change adds a new 'out_dir' option that causes the interpreter to
be invoked with the -m flag. The -m flag will point to a single
directory.