Make multiple output files in nested directories work
Right now we assume that what needs to be provided to the -m flag is the
directory name of the first output file. But it's completely valid for
output files to be spread out across directories.

This change extends this logic by computing the leading part of the
directory name that is shared by all output files.

Fixes: #56